Add js response

+++ action_ajax_form.php
@@ -6,18 +6,41 @@ $lastName = htmlspecialchars($_POST["last-name"]);
 $roleSelect = htmlspecialchars($_POST["roleSelect"]);
 $uId = htmlspecialchars($_POST["u-id"]);
 
-if(isset($_POST["delAction"]) && $_POST["delAction"] == true ){
 
+if(isset($_POST["delAction"]) && $_POST["delAction"] == true ){
 		$uIdDel = htmlspecialchars($_POST["del-u-id"]);
+
+		$sql = "SELECT * FROM users WHERE `id` = '$uIdDel'";
+		$uCount = $dbConnect->rowCount($sql);
+
+		if($uCount == 0){
+
+			$result = [
+				'status' => false, 
+				'error' => [
+					'code' => 110,
+					'message' => 'User is not found',
+				]
+			]; 
+
+			echo json_encode($result); 
+			exit;
+		}
+
 		$sql = "DELETE FROM `users` WHERE `id` = '$uIdDel' ";
 		$dbConnect->query($sql);
 
-		// Формируем массив для JSON ответа
-	    $result = [
-	    	'status' => true,
-	    	'action' => 'delete',
-	    	'uId' => $uIdDel,
-	    ]; 
+	    // Формируем массив для JSON ответа
+		$result = [
+			'status' => true,
+			'error' => null,
+			'action' => 'delete',
+			'uCount' => $uCount,
+			'user' =>
+				[
+					'uId' => $uIdDel,
+				]
+		]; 
 
 	    // Переводим массив в JSON
 	    echo json_encode($result); 
@@ -39,32 +62,58 @@ if (!empty($firstName)  && !empty($lastName) && $roleSelect != 0) {
 		$last_id = $dbConnect->insert($sql);
 		// Формируем массив для JSON ответа
 		$result = [
-		    	'firstName' => $firstName,
-		    	'lastName' => $lastName,
-		    	'status' => true,
-		    	'toggle' => $toggle,
-		    	'roleSelect' => $roleSelect,
-		    	'uId' => $last_id,
-		    	'action' => 'add',
+				'status' => true,
+				'error' => null,
+				'action' => 'add',
+				'user' =>
+		    		[
+		    			'uId' => $last_id,
+		    			'firstName' => $firstName,
+		    			'lastName' => $lastName,
+		    			'toggle' => $toggle,
+		    			'roleSelect' => $roleSelect,
+		    		]
 		    ]; 
 
 		 // Переводим массив в JSON
 		 echo json_encode($result); 
 	}else{
 
+		$sql = "SELECT * FROM users WHERE `id` = '$uId'";
+		$uCount = $dbConnect->rowCount($sql);
+
+		if($uCount == 0){
+
+			$result = [
+				'status' => false, 
+				'error' => [
+					'code' => 111,
+					'message' => 'User is not found',
+				]
+			]; 
+
+			echo json_encode($result); 
+			exit;
+		}
+
 		$sql = "UPDATE `users` SET `first_name` = '$firstName', `last_name` = '$lastName', `status` = '$toggle', `role` = '$roleSelect' WHERE `id` = '$uId' ";
 		$dbConnect->query($sql);
 
 		// Формируем массив для JSON ответа
 		$result = [
-		    	'uId' => $uId,
-		    	'firstName' => $firstName,
-		    	'lastName' => $lastName,
-		    	'status' => true,
-		    	'toggle' => $toggle,
-		    	'roleSelect' => $roleSelect,
-		    	'action' => 'edit',
-			]; 
+				'status' => true,
+				'error' => null,
+				'action' => 'edit',
+				'user' =>
+		    		[
+		    			'uId' => $uId,
+		    			'firstName' => $firstName,
+		    			'lastName' => $lastName,
+		    			'status' => true,
+		    			'toggle' => $toggle,
+		    			'roleSelect' => $roleSelect,
+		    		]
+		    ]; 
 
 		// Переводим массив в JSON
 		echo json_encode($result); 			
@@ -72,8 +121,20 @@ if (!empty($firstName)  && !empty($lastName) && $roleSelect != 0) {
 
 }else{
 
+if($roleSelect == 0){
+	$message = 'Role must be selected!';
+}else{
+	$message = 'Fields must not be empty!';	
+}
 // Формируем массив для JSON ответа
-$result = ['status' => false ]; 
+$result = [
+	'status' => false, 
+	'error' => [
+		'code' => 100,
+		'message' => $message,
+		'role' => $roleSelect,
+	]
+]; 
 
 // Переводим массив в JSON
 echo json_encode($result);
